Remove need for `unowned_workspace` fixture
This commit changes private_workspace_factory and
public_workspace_factory to use FuzzyAttributes when assigning
an owner to the workspace. This ensures that the `user` fixture
is not automatically assigned as the owner, which previously necessitated
the need for the `unowned_workspace` fixture. Since our goal is to avoid
unnecessary fixtures.