Create endpoint to retrieve current user's permission
Creates a new API endpoint, /api/workspaces/{workspace}/permissions/me/
to retreive the current permission level for the logged in user on
{workspace}. This is helpful to client applications so they can enable/disable
UI or UX elements depending on the current user's permission level for a given
workspace. If the user has explicit permission, a string representation will be
returned. If the user has no permission, but the workspace is public, a string
representation of the 'reader' permission will be returned.